Add owner-based type signature for bind.
Notice that this type signature isn't complete. The problem
is that the return type of bind() depends on the number
of arguments passed to it. e.g., if 'f' is a 3-argument function, and you call
it with 2 arguments, then it returns
a 1-argument function.

But this gets us most of the way there.

R=johnlenz
DELTA=65  (44 added, 0 deleted, 21 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3992


git-svn-id: https://closure-compiler.googlecode.com/svn/trunk@1687 b0f006be-c8cd-11de-a2e8-8d36a3108c74