Reroll RemoveUnusedClassProperties patch.
Closure style code often defines value on the prototype, we want to remove those if it is the only other property set. Previously we would leave them and the RemoveUnusedPrototypeProperties pass would remove them, but now that we would otherwise back off in the face of them we want to remove them.

Additionally, if a property is exported by convention, don't remove it in any case.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=49787523