Move the AngularJS externs files back into the Closure Compiler source tree.
While it's generally recommended to maintain externs together with the original
library, in the case of AngularJS this didn't really work: changes to the
externs are prone to break things, and we can only test those within Google.
There also seems to be little interest in maintenance of these outside of
Google, so most contributions come from inside and are much easier for users if
the source is maintained internally.

See also the corresponding AngularJS pull requests:
https://github.com/angular/angular.js/pull/5906 (v1.2/master)
https://github.com/angular/angular.js/pull/5907 (v1.0.x)
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=60021447