Handle "@suppress {const}" on var nodes ("var x = y") as well as assign nodes ("x = y").
When there is a duplicate "var" node, the Normalizer converts it to an assignment node, dropping the JsDoc in the process. So to make this work in all cases, we also need to copy the JsDoc over in Normalize.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=62212978