Add support for goog.module and updated goog.addDependency to compiler's deps scanner.
Note: this is a bit of a chicken and egg problem as goog.module tests won't run without some deps support for it.  Tools need to be updated to accept and generate an extra value for goog.addDependency so that the file is loaded the correct way.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=68540366