Updated readme to point to the releases wiki page
When people visit the project page to download the latest, they want to know what the latest version is! ("Has it changed since I was last here?").

Even if you download the latest you have no easy way to tell what version it is - nothing in the zip mentions current version.

I propose adding a link to the release wiki page, which is kinda buried otherwise, since there's no ToC in the wiki.