Automated rollback.
*** Original change description ***

Fix transpilation of for/of loops where the iteration variable is declared with let or const.

In this case, it may shadow a variable from the outer scope, so we should keep it as a let or const, so that the Es6RewriteLetConst pass can rename it if needed.

Fixes #717 on Github.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=82976608