Objects added to array via goog.array.extend
Eventful objects added to an array are not tracked further in this pass.
goog.array.extend is another way of adding objects to an array. Treat calls to
goog.array.extend as if it is disposing all arguments as the first argument is
an array and there are a variable number of arguments that could be added.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=84474637