Split StaticSlot, StaticReference and StaticScope into typed and untyped versions. Make StaticSymbolTable untyped. Make StaticSourceFile untyped. Move untyped interfaces out of the jstype package.
Context: the StaticSymbolTable interface is not used by the compiler. It is only used by the SymbolTable class, which creates a symbol table for the compiler's custom clients. By making StaticSymbolTable untyped, we allow the 4 classes that implement the interface to use either typed or untyped scopes. In the next CL, I will split Scope and Var into typed and untyped versions (the original goal), and change several classes in the compiler to use untyped scopes and vars.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=87348168