Run the check for require/provide sorting before goog.module processing.
This way, if the requires are sorted by alias instead of full namespace, we won't warn about it. Eventually, we'll want to add to the check so that it asserts the requires *are* sorted by alias name.

As an example, the check currently expects

var beta = goog.require('alpha.beta');
var alpha = goog.require('beta.alpha');

to be in that order, but they should actually be in the reverse order.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=101368621