closure: update link and comment in Google Analytics externs file
Documentation of the ga.js API has migrated to new location so update
the link in the file.  Furthermore, ga.js API is now called a ?Classic
Analytics? (as opposed to ?Universal Analytics? which is the new thing)
so update comment to reflact that so there is no confusion.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=101601400