Add support for computed getters and setters in ES6 classes.
At least for now, the key must be a qualified name. This simplifies things in the compiler because we don't have to worry about side effects when the key is evaluated. But I think this restriction will be okay, because the only thing we expect people to use as a key is a qualified name which refers to a Symbol.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=102098974