Change usage of "*" to represent the unknown object in a Promise rejection to using "?".
"*" represents the top type in JSCompiler, whereas "?" represents the "any" type, i.e. an expression that cannot be typed and should not be checked.

The goal is to make arbitrary handler functions proper subtypes of the required function type, so that user code can (sort of cheatingly) pretend that the function type was bivariant.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=109449847