Back off on the @param checks for functions that aren't in the global scope.
If there are any @param annotations or inline param annotations, we still check all the params. But if not, the presence of another annotation (such as @ngInject) does not trigger these checks, unless the function is in the global scope.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=113080022