Don't warn about missing requires on namespaces defined in the same file.
Adding a goog.require wouldn't be a way of fixing these cases anyway.

This is particularly important as we switch to code using modules.
e.g. It doesn't make sense to ask for a require of FooBar.Subclass
for the following code:
    function FooBar() {};
    FooBar.Subclass = makeConstructor();
    new FooBar.Subclass();
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116063490