Requires check: Make sure we warn about classes like goog.Promise and goog.Timer
We are already making an exception where we don't warn for several "goog.*" namespaces. This is to avoid things like goog.isDef which are defined in base.js and thus do not need to be required. However, there are several classes (goog.Promise, goog.Timer, and so on) which were erroneously caught by this exception. If the second part of the string looks like a class name, warn about the missing require.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=118600416