Attach JS coverage data to window.top.__jscov.
The coverage data extractor injects a script. To support parallel testing
(tests in iframes) and module testing (src enclosed in modules) we need
the extractor-script and data in the same scope. The simple way to
accomplish this is to attach all of the data at the tippy-top of global,
window.top.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=121599469