Hoists constant like fields in J2CL generated code.
Some fields in Java are not compile time constants but eventually become assigned by a constant value after JsCompiler optimizations (e.g. System.getProperty calls and other operations that are statically evaluate-able).
As such fields are not compile time constants, they are both assigned in module scope as part of class declaration and also inside $clinit for initialization. As a result, JsCompiler refuses to inline and statically evaluate expressions that refers to field because it doesn't know when it will get initialized.

Technically, this is correct behavior since there are no guarantees that the initial value will be visible by others or not. However, such value is only visible in Java if there is a cycle in clinits and the field is accessed before initialization. That is almost always a bug and GWT never assumed this state is observable in its optimizations, yet nobody complained. Hence we consider it is safe to upgrade such field to a constant and make it inline-able as GWT does.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=121626772