[NTI] Specialize the lhs type appropriately when handling logical ops.
Because of short-circuiting, we know that an OR expression will
never evaluate to any falsy values its lhs might take, so we can
specialize the lhs type to truthy before joining with the rhs type.
Similarly for an AND we can specialize the lhs type to falsy.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=123980865