Add Array.prototype.includes to es6.js externs.
Note that this is technically in ES2016 (= ES7) but this is the only
new function added in ES2016 and it is making less and less sense to
split these out by versions.  I have modified the comment at the top
of the file to indicate that it includes ES6 *and later* externs, but
at some point the core externs should ideally be restructured.

RELNOTES: n/a
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124793921