Fix source map generator to produce valid JSON.
Currently it escapes several sequences like "-->", "]]>", "<!--" and "</script" by inserting a backslash in the middle, but this causes Chrome's source map parser to fail to decode the JSON since "\>", "\!" and "\/" are not valid JSON escapes.  I fix this by using a unicode escape for one of the relevant characters, which still prevents the characters from being misinterpreted as comments or script/cdata tags, but allows the JSON parser to handle it correctly.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=127386068