Allow goog.module.get of goog.provided name
This makes goog.module.get fully agnostic to whether it's target is a
goog.module or a goog.provide, to ease transition to goog.module.

As a side effect, also be more more strict about goog.module.get depending
on a non-existent file. This used to be allowed for some reason, but is now
forbidden.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=127969546